Remove the `SyncJob` interface
