+++ demo/src/main/java/com/google/android/fhir/demo/MainActivityViewModel.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,23 +17,20 @@
 package com.google.android.fhir.demo
 
 import android.app.Application
-import android.text.format.DateFormat
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
 import androidx.work.Constraints
-import com.google.android.fhir.demo.data.FhirPeriodicSyncWorker
+import com.google.android.fhir.demo.data.FhirSyncWorker
 import com.google.android.fhir.sync.PeriodicSyncConfiguration
 import com.google.android.fhir.sync.RepeatInterval
 import com.google.android.fhir.sync.State
 import com.google.android.fhir.sync.Sync
-import java.time.format.DateTimeFormatter
 import java.util.concurrent.TimeUnit
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
 /** View model for [MainActivity]. */
@@ -43,41 +40,20 @@ class MainActivityViewModel(application: Application, private val state: SavedSt
   val lastSyncTimestampLiveData: LiveData<String>
     get() = _lastSyncTimestampLiveData
 
-  private val job = Sync.basicSyncJob(application.applicationContext)
   private val _pollState = MutableSharedFlow<State>()
   val pollState: Flow<State>
     get() = _pollState
 
   init {
-    poll()
-  }
-
-  /** Requests periodic sync. */
-  fun poll() {
     viewModelScope.launch {
-      job.poll(
+      Sync.periodicSync<FhirSyncWorker>(
+          application.applicationContext,
           PeriodicSyncConfiguration(
             syncConstraints = Constraints.Builder().build(),
             repeat = RepeatInterval(interval = 15, timeUnit = TimeUnit.MINUTES)
-          ),
-          FhirPeriodicSyncWorker::class.java
+          )
         )
         .collect { _pollState.emit(it) }
     }
-  }
-
-  /** Emits last sync time. */
-  fun updateLastSyncTimestamp() {
-    val formatter =
-      DateTimeFormatter.ofPattern(
-        if (DateFormat.is24HourFormat(getApplication())) formatString24 else formatString12
-      )
-    _lastSyncTimestampLiveData.value =
-      job.lastSyncTimestamp()?.toLocalDateTime()?.format(formatter) ?: ""
-  }
-
-  companion object {
-    private const val formatString24 = "yyyy-MM-dd HH:mm:ss"
-    private const val formatString12 = "yyyy-MM-dd hh:mm:ss a"
   }
 }
